+++ shipwright/exceptions.py
@@ -1,0 +1,5 @@
+from __future__ import absolute_import
+
+from ._lib.source_control import SourceControlNotFound
+
+__all__ = ['SourceControlNotFound']