+++ shipwright/targets.py
@@ -1,0 +1,6 @@
+from __future__ import absolute_import
+
+from ._lib.source_control import AUTO, GIT
+from ._lib.targets import targets
+
+__all__ = ['targets', 'AUTO', 'GIT']